PC Instagram Hyperlink
An executable that opens the Instagram webpage on default browser, and changes to mobile mode.
Originally designed just for self-use.
The executable includes java.awt.Robot component, which performs keyboard clicks automatically, and the entire program only runs for 2 - 3 seconds, then it will terminate automatically.
The program cannot guaranteed 100% success rate even on suitable browsers. Try running the program for several times if it doesn't work.
The code is also presented in this github. The code is malware-free.